+++ src/services/helpers/client.rs
@@ -1,6 +1,6 @@
 use crate::prelude::*;
 
-use crate::services::helpers::middleware::{inject_default_headers, timeout_request};
+use crate::services::helpers::middleware::{error_for_status, inject_default_headers, timeout_request};
 use lazy_static::lazy_static;
 use surf::middleware::Redirect;
 
@@ -9,5 +9,6 @@ lazy_static! {
     pub static ref CLIENT: Client = surf::client()
         .with(Redirect::default())
         .with(inject_default_headers)
-        .with(timeout_request);
+        .with(timeout_request)
+        .with(error_for_status);
 }